[analyzer] Attempt to make a windows buildbot happy.
Got an error that a cast is happening from a pointer type to long, which is
smaller.

llvm-svn: 342223

+++ clang/test/Analysis/cxx-uninitialized-object-ptr-ref.cpp
@@ -25,7 +25,7 @@ void fConcreteIntLocTest() {
 // nonloc::LocAsInteger tests.
 //===----------------------------------------------------------------------===//
 
-using intptr_t = long;
+using intptr_t = unsigned long long;
 
 struct LocAsIntegerTest {
   intptr_t ptr; // expected-note{{uninitialized pointee 'reinterpret_cast<char *>(this->ptr)'}}
